ECA-2725: CAInfo.setValidity should have long parameter
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@16424 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ modules/ejbca-ejb-cli/src/org/ejbca/ui/cli/ca/CaInitCommand.java
@@ -224,7 +224,7 @@ public class CaInitCommand extends BaseCaAdminCommand {
             }
             final String keyspec = args[5];
             final String keytype = args[6];
-            final int validity = Integer.parseInt(args[7]);
+            final long validity = Long.parseLong(args[7]);
             String policyId = args[8];
             final ArrayList<CertificatePolicy> policies = new ArrayList<CertificatePolicy>(1);
             if ((policyId != null) && (policyId.toLowerCase().trim().equals("null"))) {
